Refactor: Simplify HomePage and WorkFlow components; remove unused imports and automation templates; implement AutomationAiSupport component for better organization.
